refactor: add base model interface to get models id and change responsable attribute to be an User instead of a string
