+++ src/api/domain/models/base-model.ts
@@ -1,0 +1,3 @@
+export interface BaseModel{
+    getId(): number | string | undefined;
+}
