+++ src/api/domain/models/user.ts
@@ -1,7 +1,17 @@
-export default class User {
-    constructor(private email: string){}
+import { BaseModel } from "./base-model";
+
+export default class User implements BaseModel{
+    constructor(
+        private id?: string,
+        private email?: string
+    ){}
+
+    // Override
+    getId(): string | undefined {
+        return this.id
+    }
     
-    getEmail(): string{
+    getEmail(): string | undefined{
         return this.email
     }
